+++ lib/src/hooks/use_form_context.dart
@@ -12,7 +12,7 @@ class _HookedFormProviderBase extends InheritedWidget {
 
   final FormFieldsController form;
 
-  static FormFieldsController<F> of<F extends FieldSchema<dynamic>>(
+  static FormFieldsController<FieldSchema> of(
     BuildContext context,
   ) {
     final widget = context
@@ -29,7 +29,7 @@ class _HookedFormProviderBase extends InheritedWidget {
       );
     }
 
-    return widget.form as FormFieldsController<F>;
+    return widget.form;
   }
 
   @override
@@ -49,7 +49,8 @@ class _HookedFormProviderBase extends InheritedWidget {
 ///   child: MyFormWidget(),
 /// )
 /// ```
-class HookedFormProvider<F extends FieldSchema<dynamic>> extends StatelessWidget {
+class HookedFormProvider<F extends FieldSchema<dynamic>>
+    extends StatelessWidget {
   /// Creates a [HookedFormProvider] that provides a [FormFieldsController] to the form fields.
   const HookedFormProvider({
     super.key,
@@ -86,8 +87,8 @@ class HookedFormProvider<F extends FieldSchema<dynamic>> extends StatelessWidget
 ///
 /// DO NOT use this hook to create a [FormFieldsController], please see
 /// [useForm] instead.
-FormFieldsController<F> useFormContext<F extends FieldSchema<dynamic>>(
+FormFieldsController<FieldSchema> useFormContext(
   BuildContext context,
 ) {
-  return _HookedFormProviderBase.of<F>(context);
+  return _HookedFormProviderBase.of(context);
 }